+++ packages/everything-web/packages/react-sdk/src/index.ts
@@ -10,6 +10,8 @@ export {
 export { chunkElements, getModeAspectRatio, calculateLayoutSizes } from './utils/layout';
 export { usePreview } from './hooks/usePreview'
 export { useVideoTile } from './hooks/useVideoTile'
-export { useVideoList } from './hooks/useVideoListLayout'
+export { useVideoList } from './hooks/useVideoList'
 export { useAVToggle } from './hooks/useAVToggle'
 export { useDevices } from './hooks/useDevices'
+export { useVideo } from './hooks/useVideo'
+export { useAudioLevel } from './hooks/useAudioLevel'