+++ packages/everything-web/packages/react-sdk/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2-alpha
+
+-   `useVideo` hook
+-   `useAudioLevel` hook
+-   make `showScreenFn` optional in `useVideoList`
+-   make `userName` optional in `usePreview`
+
 ## 0.0.1-alpha
 
 -   `useVideoList()` take config as an object instead of params